Move RangeSlider to new components
